Moved Submission Tests to another folder
The submission tests are getting in the way of my own testing and it's becoming hard to figure out my own testing.
