+++ .gitignore
@@ -7,6 +7,7 @@ node_modules/
 
 # testing
 coverage/
+.nyc_output/
 
 # production
 build/
