+++ test/submission-tests/caesar.submission.test.js
@@ -1,5 +1,5 @@
 const { expect } = require("chai");
-const { caesar } = require("../src/caesar");
+const { caesar } = require("../../src/caesar");
 
 describe("caesar() submission tests written by Thinkful", () => {
   describe("error handling", () => {
