+++ test/submission-tests/polybius.submission.test.js
@@ -1,5 +1,5 @@
 const { expect } = require("chai");
-const { polybius } = require("../src/polybius");
+const { polybius } = require("../../src/polybius");
 
 describe("polybius() submission tests written by Thinkful", () => {
   describe("encoding a message", () => {
